better cast in set_dac()

+++ ldk/drivers/oscillo.py
@@ -109,8 +109,8 @@ class Oscillo(object):
         def set_dac_buffer(self, channel, arr):
             pass
         for channel in channels:
-            data = np.uint32(16384 * (self.dac[channel,:]))
-            set_dac_buffer(self, channel, data[::2] + data[1::2] * 65536)
+            data = np.int16(16384 * (self.dac[channel,:]))
+            set_dac_buffer(self, channel, np.uint32(data[::2] + data[1::2] * 65536))
 
     @command()
     def set_averaging(self, avg_status):
